requirements.txt
Django==3.1.2
selenium==3.141.0

djangorestframework~=3.12.2
Pygments~=2.8.1